Tweak the nudging factor for markers, and better UI for markers.
These changes are to improve AnimatedIcon workflow.
The nudging value of 0.01 was insufficient, so this ups it to 0.05.
The UI in LottieViewer for markers and the scrubber is confusing because it shows progress, which designers don't care about. So now it uses frames everywhere. Developers that need to convert to progress can do the math themselves (it's just (frame count) / (number of frames) plus a nudge.

+++ source/UIDataCodeGen/CodeGen/Tables/LottieMarkersMonospaceTableFormatter.cs
@@ -91,7 +91,7 @@ namespace Microsoft.Toolkit.Uwp.UI.Lottie.UIData.CodeGen.Tables
                     ColumnData.Create(m.StartTime.TotalMilliseconds),
                     m.DurationInFrames > 0 ? ColumnData.Create(m.Duration.TotalMilliseconds) : ColumnData.Empty,
                     ColumnData.Create(stringifier.Float(m.StartProgress), TextAlignment.Left),
-                    m.DurationInFrames > 0 ? ColumnData.Create(stringifier.Float(m.EndProgress), TextAlignment.Left) : ColumnData.Empty
+                    m.DurationInFrames > 0 ? ColumnData.Create(stringifier.Double(m.EndProgress), TextAlignment.Left) : ColumnData.Empty
                 );
 
             records = records.Append(Row.BodyBottom);
